{{cookiecutter.repo_name}}/requirements/base.txt

redis
appdirs
olefile
packaging
Pillow==4.0.0
pyparsing==2.1.10
rcssmin
rjsmin
six


# Django
Django=={{cookiecutter.django_version}}
# PROTECTED REGION ID({{cookiecutter.project.name}}.django.requirements) ENABLED START
django-classy-tags
django-reversion
django-rq
django-compressor
django-sekizai
# PROTECTED REGION END

{% if cookiecutter.django_cms_version %}
# Django CMS
django-cms=={{cookiecutter.django_cms_version}}
# PROTECTED REGION ID({{cookiecutter.project.name}}.django_cms.requirements) ENABLED START
django-appconf
#django-formtools
django-treebeard
djangocms-admin-style
djangocms-text-ckeditor
djangocms-style
djangocms-googlemap
djangocms-inherit
Django-Select2
django-categories
# Django filer plugin
django-filer
easy-thumbnails
cmsplugin-filer
# PROTECTED REGION END
{% endif %}

{% if cookiecutter.django_rest_framework_version %}
djangorestframework=={{cookiecutter.django_rest_framework_version}}
# PROTECTED REGION ID({{cookiecutter.project.name}}.django_rest_framework.requirements) ENABLED START
# PROTECTED REGION END
{% endif %}

{% if cookiecutter.zappa == 'y' %}
zappa
# PROTECTED REGION ID({{cookiecutter.project.name}}.zappa.requirements) ENABLED START
# PROTECTED REGION END
{% endif %}
